+++ dwolla-sdk/src/main/scala/dwolla/sdk/DwollaApi.scala
@@ -17,9 +17,6 @@ private[sdk] trait DwollaApi {
   def addFundingSource(accessToken: String, accountNumber: String, routingNumber: String, accountType: AccountType,
                        name: String): Future[AddFundingSourceResponse]
 
-  def addFundingSource(accessToken: String, accountNumber: String, routingNumber: String, accountType: String,
-                       name: String): Future[AddFundingSourceResponse]
-
   def getFundingSourceDetails(accessToken: String, fundingId: String): Future[GetFundingSourceDetailsResponse]
 
   def depositFunds(accessToken: String, fundingId: String, pin: String,