+++ dwolla-sdk/src/main/scala/dwolla/sdk/DwollaApiRequestJsonProtocol.scala
@@ -22,14 +22,12 @@ private[sdk] object DwollaApiRequestJsonProtocol extends DefaultJsonProtocol {
 
   implicit val accountTypeFormat = new JsonFormat[AccountType] {
     def write(accountType: AccountType) = accountType match {
-      case AccountType.Checking => JsString("Checking")
-      case AccountType.Savings => JsString("Savings")
+      case Requests.AccountType.Checking => JsString("Checking")
+      case Requests.AccountType.Savings => JsString("Savings")
     }
 
     def read(value: JsValue) = ???
   }
-
-  implicit val addFundingSourceRequest2Format = jsonFormat5(AddFundingSourceRequest2)
 
   implicit val addFundingSourceRequestFormat = jsonFormat5(AddFundingSourceRequest)