+++ dwolla-sdk/src/main/scala/dwolla/sdk/DwollaSdk.scala
@@ -144,15 +144,6 @@ class DwollaSdk(settings: Option[DwollaApiSettings] = None)(
       } yield addFundingSourceResponse
     }
 
-    @deprecated("Use the DwollaSdk.FundingSource.create overload which provides an AccountType parameter", "1.1.1")
-    def create(accessToken: String, accountNumber: String, routingNumber: String, accountType: String,
-               name: String): Future[FundingSource] = {
-      for {
-        addFundingSourceResponse <- dwollaApi.addFundingSource(accessToken, accountNumber, routingNumber,
-          accountType, name)
-      } yield addFundingSourceResponse
-    }
-
     def retrieve(accessToken: String, id: String): Future[FundingSource] = {
       for {
         getFundingSourceDetailsResponse <- dwollaApi.getFundingSourceDetails(accessToken, id)